refactor: Remove Comment component and update layout in Parameters and DetectorImage components for improved responsiveness
